+++ WeatherApp/Extentions/UIImageView.swift
@@ -1,0 +1,24 @@
+//
+//  UIImageView.swift
+//  WeatherApp
+//
+//  Created by Munsif Hayat on 03/02/2022.
+//
+
+import UIKit
+
+extension UIImageView {
+    func load(url: URL) {
+        DispatchQueue.global().async { [weak self] in
+            if let data = try? Data(contentsOf: url) {
+                if let image = UIImage(data: data) {
+                    DispatchQueue.main.async {
+                        self?.image = image
+                    }
+                }
+            } else {
+                print("Nothing")
+            }
+        }
+    }
+}
